Consolidate single-value constants
